revert: Changes Session back to own the socket
This was done as a fix for the server, but it seems that letting T and simply passing &mut T fixes the original issue this PR was intended to fix.

Other libraries like embedded_tls take ownership of the socket. This allows to move session around without having to keep a reference to the socket alive.

Ref: c094cfc